Add: Offset

+++ pci/src/xhci/registers/operation_registers/usb_command_register/host_controller_reset.rs
@@ -1,4 +1,6 @@
 use macros::VolatileFlag;
 
+use crate::error::{PciError, PciResult};
+
 #[derive(VolatileFlag)]
-pub struct HostControllerReset(u8);
+pub struct HostControllerReset(usize);
